+++ info/endpoints/endpoints.AccountTaxonomy.27bb991a.js
@@ -1,10 +1,10 @@
 (self.webpackChunk_twitter_responsive_web =
   self.webpackChunk_twitter_responsive_web || []).push([
-  ["endpoints.Graphql"],
+  ["endpoints.AccountTaxonomy", "endpoints.Graphql"],
   {
     29848: (e) => {
       e.exports = {
-        queryId: "VeIlDt0K1vcCzf-jW4qd6A",
+        queryId: "h52d1F7dumWGE1tJAhQBpg",
         operationName: "BlockedAccountsAll",
         operationType: "query",
         metadata: {
@@ -34,7 +34,7 @@
     },
     57274: (e) => {
       e.exports = {
-        queryId: "h9d1UQkdRub8jR5XVY_uuQ",
+        queryId: "8w-D2OhT0jmGzXaNY--UQA",
         operationName: "BlockedAccountsAutoBlock",
         operationType: "query",
         metadata: {
@@ -64,7 +64,7 @@
     },
     2216: (e) => {
       e.exports = {
-        queryId: "gZpIitopgjdaqdwK6reCtg",
+        queryId: "8LDNeOEm0kA98uoDsqXvMg",
         operationName: "BlockedAccountsImported",
         operationType: "query",
         metadata: {
@@ -92,9 +92,17 @@
         },
       };
     },
+    29297: (e) => {
+      e.exports = {
+        queryId: "_ckHEj05gan2VfNHG6thBA",
+        operationName: "DisableUserAccountLabel",
+        operationType: "mutation",
+        metadata: { featureSwitches: [] },
+      };
+    },
     41839: (e) => {
       e.exports = {
-        queryId: "oW2F-EMg_7rtTJgDShBe5w",
+        queryId: "JpFFCTBPxYVlDqMUr9twzQ",
         operationName: "Followers",
         operationType: "query",
         metadata: {
@@ -124,7 +132,7 @@
     },
     97621: (e) => {
       e.exports = {
-        queryId: "1go5V_4KX2r8NMJgmPkhww",
+        queryId: "RvojYJJB90VwJ0rdVhbjMQ",
         operationName: "FollowersYouKnow",
         operationType: "query",
         metadata: {
@@ -154,7 +162,7 @@
     },
     87669: (e) => {
       e.exports = {
-        queryId: "RsIOCb3LhS_AzOkBxJiJIg",
+        queryId: "b22I8WSfQ8H4Ev8486xAlQ",
         operationName: "Following",
         operationType: "query",
         metadata: {
@@ -184,7 +192,7 @@
     },
     75840: (e) => {
       e.exports = {
-        queryId: "bnQ1HNHhBA4S_iB1SaT9HA",
+        queryId: "LZfAdxTdNolKXw6ZkoY_kA",
         operationName: "GenericTimelineById",
         operationType: "query",
         metadata: {
@@ -214,7 +222,7 @@
     },
     50931: (e) => {
       e.exports = {
-        queryId: "p3kCiX2fbX2lRjYgo6h6Ew",
+        queryId: "hnaqw2Vok5OETdBVa_uexw",
         operationName: "ModeratedTimeline",
         operationType: "query",
         metadata: {
@@ -244,7 +252,7 @@
     },
     59113: (e) => {
       e.exports = {
-        queryId: "003287_lTwFeSs_lOSw-HA",
+        queryId: "-G9eXTmseyiSenbqjrEG6w",
         operationName: "MutedAccounts",
         operationType: "query",
         metadata: {
@@ -274,7 +282,7 @@
     },
     26491: (e) => {
       e.exports = {
-        queryId: "-UBJTrL7wLKQXV-x9-bVZg",
+        queryId: "o0YtPFnd4Lk_pOQb9alCvA",
         operationName: "SuperFollowers",
         operationType: "query",
         metadata: {
@@ -302,6 +310,14 @@
         },
       };
     },
+    51659: (e) => {
+      e.exports = {
+        queryId: "rD5gLxVmMvtdtYU1UHWlFQ",
+        operationName: "UserAccountLabel",
+        operationType: "query",
+        metadata: { featureSwitches: [] },
+      };
+    },
     16221: (e) => {
       e.exports = {
         queryId: "D8mVcJSVv66_3NcR7fOf6g",
@@ -318,24 +334,64 @@
         },
       };
     },
+    95901: (e, _, t) => {
+      "use strict";
+      t.r(_),
+        t.d(_, {
+          default: () => b,
+          isDisableUserAccountLabelFail: () => p,
+          isFetchFail: () => d,
+        });
+      var n = t(72599),
+        i = t(17360),
+        a = t(29297),
+        r = t.n(a),
+        l = t(51659),
+        o = t.n(l),
+        s = t(97534);
+      const d = (e, _) => {
+          var t;
+          return (
+            (null == _ || null == (t = _.user) ? void 0 : t.result) ||
+              (0, n.ZP)(
+                "GQL AccountTaxonomy: Failed to load User Account Label",
+                { extra: (0, s.errorBreadcrumb)(e) }
+              ),
+            (0, i.jB)(e)
+          );
+        },
+        p = (e, _) => (
+          (null == _ ? void 0 : _.disable_user_account_label) ||
+            (0, n.ZP)(
+              "GQL AccountTaxonomy: Failed to load Disable User Account Status",
+              { extra: (0, s.errorBreadcrumb)(e) }
+            ),
+          (0, i.jB)(e)
+        ),
+        b = ({ apiClient: e, featureSwitches: _ }) => ({
+          fetchUserAccountLabel: ({ userId: _ }) =>
+            e.graphQL(o(), { rest_id: _ }, d),
+          disableUserAccountLabel: () => e.graphQL(r(), {}, p),
+        });
+    },
     97534: (e, _, t) => {
       "use strict";
       t.r(_),
         t.d(_, {
-          default: () => V,
-          errorBreadcrumb: () => C,
-          isFatalGenericTimelineError: () => H,
-          isFatalModeratedTimelineError: () => Y,
-          isFatalTeamsResponseError: () => U,
-          isFatalUserScopedFollowTimelineError: () => E,
-          isFatalViewerScopedFollowTimelineError: () => N,
-          parseUserScopedTimeline: () => Z,
+          default: () => M,
+          errorBreadcrumb: () => Z,
+          isFatalGenericTimelineError: () => V,
+          isFatalModeratedTimelineError: () => j,
+          isFatalTeamsResponseError: () => E,
+          isFatalUserScopedFollowTimelineError: () => N,
+          isFatalViewerScopedFollowTimelineError: () => C,
+          parseUserScopedTimeline: () => z,
           parseViewerScopedTimeline: () => P,
         });
       var n = t(58955),
         i = t(79363),
-        r = t(72599),
-        a = t(32091),
+        a = t(72599),
+        r = t(32091),
         l = t(17360),
         o = t(83175),
         s = t(6899),
@@ -362,25 +418,25 @@
         m = t(2216),
         h = t.n(m),
         v = t(41839),
-        g = t.n(v),
-        f = t(97621),
-        y = t.n(f),
+        f = t.n(v),
+        g = t(97621),
+        y = t.n(g),
         q = t(87669),
         x = t.n(q),
-        k = t(75840),
-        T = t.n(k),
-        A = t(50931),
-        I = t.n(A),
-        B = t(59113),
-        S = t.n(B),
-        F = t(26491),
-        G = t.n(F),
-        z = t(16221),
-        Q = t.n(z),
-        L = t(82249);
-      const Z = (e) => {
+        A = t(75840),
+        k = t.n(A),
+        T = t(50931),
+        B = t.n(T),
+        F = t(59113),
+        I = t.n(F),
+        L = t(26491),
+        G = t.n(L),
+        Q = t(16221),
+        S = t.n(Q),
+        U = t(82249);
+      const z = (e) => {
           var _;
-          let t = L.cY;
+          let t = U.cY;
           var n;
           null != e &&
             null != (_ = e.user) &&
@@ -388,23 +444,23 @@
             "User" === e.user.result.__typename &&
             (t =
               (null == (n = e.user.result.timeline) ? void 0 : n.timeline) ||
-              L.cY);
+              U.cY);
           return t;
         },
-        C = (e) => {
+        Z = (e) => {
           var _;
           return { code: null == (_ = e[0]) ? void 0 : _.code };
         },
         P = (e) => {
-          let _ = L.cY;
+          let _ = U.cY;
           var t;
           null != e &&
             e.viewer &&
             (_ =
-              (null == (t = e.viewer.timeline) ? void 0 : t.timeline) || L.cY);
+              (null == (t = e.viewer.timeline) ? void 0 : t.timeline) || U.cY);
           return _;
         },
-        U = (e) =>
+        E = (e) =>
           Array.isArray(e) &&
           !(0, i.Z)(
             e,
@@ -412,7 +468,7 @@
               Array.isArray(e.path) &&
               /^viewer.teams.act_as.\d+(.|$)/.test(e.path.join("."))
           ),
-        E = (e, _) => {
+        N = (e, _) => {
           var t, n;
           const i =
             "User" ===
@@ -421,24 +477,24 @@
                 : n.__typename) && _.user.result.timeline;
           return (
             i ||
-              (0, r.ZP)("GQL URT: Failed to render Follow timeline", {
-                extra: C(e),
+              (0, a.ZP)("GQL URT: Failed to render Follow timeline", {
+                extra: Z(e),
               }),
             !i && (0, l.jB)(e)
           );
         },
-        N = (e, _) => {
+        C = (e, _) => {
           var t;
           const n = null == _ || null == (t = _.viewer) ? void 0 : t.timeline;
           return (
             n ||
-              (0, r.ZP)("GQL URT: Failed to render Subscriptions timeline", {
-                extra: C(e),
+              (0, a.ZP)("GQL URT: Failed to render Subscriptions timeline", {
+                extra: Z(e),
               }),
             !n && (0, l.jB)(e)
           );
         },
-        R = (e, _) => {
+        Y = (e, _) => {
           var t, n;
           const i =
             null == _ ||
@@ -448,13 +504,13 @@
               : n.timeline;
           return (
             i ||
-              (0, r.ZP)("GQL URT: Failed to render Muting timeline", {
-                extra: C(e),
+              (0, a.ZP)("GQL URT: Failed to render Muting timeline", {
+                extra: Z(e),
               }),
             !i && (0, l.jB)(e)
           );
         },
-        Y = (e, _) => {
+        j = (e, _) => {
           var t, n;
           const i =
             "Tweet" ===
@@ -466,25 +522,25 @@
               : n.timeline);
           return (
             i ||
-              (0, r.ZP)(
+              (0, a.ZP)(
                 "GQL URT: Failed to render ModeratedTimeline timeline",
-                { extra: C(e) }
+                { extra: Z(e) }
               ),
             !i && (0, l.jB)(e)
           );
         },
-        H = (e, _) => {
+        V = (e, _) => {
           var t;
           const n = null == _ || null == (t = _.timeline) ? void 0 : t.timeline;
           return (
             n ||
-              (0, r.ZP)("GQL URT: Failed to render GenericTimelineById", {
-                extra: C(e),
+              (0, a.ZP)("GQL URT: Failed to render GenericTimelineById", {
+                extra: Z(e),
               }),
             !n && (0, l.jB)(e)
           );
         };
-      function j(e, _, t) {
+      function H(e, _, t) {
         let n;
         function i(e) {
           return (
@@ -496,12 +552,12 @@
           );
         }
         return (
-          (n = e === a.kH.Imported ? h() : e === a.kH.AutoBlock ? w() : c()),
-          ({ count: a, cursor: s }) =>
+          (n = e === r.kH.Imported ? h() : e === r.kH.AutoBlock ? w() : c()),
+          ({ count: r, cursor: s }) =>
             _.graphQL(
               n,
               {
-                count: a,
+                count: r,
                 cursor: s,
                 includePromotedContent: !1,
                 ...(0, o.d)(t),
@@ -513,24 +569,24 @@
                 const n = i(t);
                 return (
                   n ||
-                    (0, r.ZP)(
+                    (0, a.ZP)(
                       `GQL URT: Failed to query for Blocked Accounts [${e}]`,
-                      { extra: C(_) }
+                      { extra: Z(_) }
                     ),
                   !n && (0, l.jB)(_)
                 );
               }
-            ).then((e) => i(e) || L.cY)
+            ).then((e) => i(e) || U.cY)
         );
       }
-      const V = ({ apiClient: e, featureSwitches: _ }) => ({
-        fetchBlockedAccountsImportedGraphql: j(a.kH.Imported, e, _),
-        fetchBlockedAccountsAllGraphql: j(a.kH.All, e, _),
-        fetchBlockedAccountsAutoBlockGraphql: j(a.kH.AutoBlock, e, _),
+      const M = ({ apiClient: e, featureSwitches: _ }) => ({
+        fetchBlockedAccountsImportedGraphql: H(r.kH.Imported, e, _),
+        fetchBlockedAccountsAllGraphql: H(r.kH.All, e, _),
+        fetchBlockedAccountsAutoBlockGraphql: H(r.kH.AutoBlock, e, _),
         fetchFollowers: ({ count: t, cursor: n, userId: i }) =>
           e
             .graphQL(
-              g(),
+              f(),
               {
                 userId: i,
                 count: t,
@@ -538,9 +594,9 @@
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              E
+              N
             )
-            .then(Z),
+            .then(z),
         fetchFollowersYouKnow: ({ count: t, cursor: n, userId: i }) =>
           e
             .graphQL(
@@ -552,9 +608,9 @@
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              E
+              N
             )
-            .then(Z),
+            .then(z),
         fetchFollowing: ({ count: t, cursor: n, userId: i }) =>
           e
             .graphQL(
@@ -566,9 +622,9 @@
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              E
+              N
             )
-            .then(Z),
+            .then(z),
         fetchSuperFollowers: ({ count: t, cursor: n }) =>
           e
             .graphQL(
@@ -579,24 +635,24 @@
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              N
+              C
             )
             .then(P),
         fetchViewerTeams: () =>
           e
-            .graphQLFullResponse(Q(), { ...(0, o.S)(_) }, U)
+            .graphQLFullResponse(S(), { ...(0, o.S)(_) }, E)
             .then(({ data: e, errors: _ }) => {
               const { viewer: t } = e,
                 i = (t && t.teams && t.teams.act_as) || [];
               if (!_ || !_.length) return i;
-              const r = (0, n.Z)(_.map((e) => e.path && e.path[3]));
-              return i.filter((e, _) => !r.has(_));
+              const a = (0, n.Z)(_.map((e) => e.path && e.path[3]));
+              return i.filter((e, _) => !a.has(_));
             })
             .then(p),
         fetchModeratedTimeline: ({ count: t, cursor: n, rootTweetId: i }) =>
           e
             .graphQL(
-              I(),
+              B(),
               {
                 rootTweetId: i,
                 count: t,
@@ -604,7 +660,7 @@
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              Y
+              j
             )
             .then((e) => {
               var _;
@@ -617,35 +673,35 @@
                   (null == (t = e.tweet.result) ||
                   null == (n = t.timeline_response)
                     ? void 0
-                    : n.timeline) || L.cY
+                    : n.timeline) || U.cY
                 );
               }
-              return L.cY;
+              return U.cY;
             }),
         fetchMutedAccounts: ({ count: t, cursor: n }) =>
           e
             .graphQL(
-              S(),
+              I(),
               {
                 count: t,
                 cursor: n,
                 includePromotedContent: !1,
                 ...(0, o.d)(_),
               },
-              R
+              Y
             )
             .then((e) => {
               var _, t;
               return (
                 (null == (_ = e.viewer) || null == (t = _.muting_timeline)
                   ? void 0
-                  : t.timeline) || L.cY
+                  : t.timeline) || U.cY
               );
             }),
         fetchGenericTimelineById: ({ count: t, cursor: n, timelineId: i }) =>
           e
             .graphQL(
-              T(),
+              k(),
               {
                 timelineId: i,
                 count: t,
@@ -653,40 +709,40 @@
                 withQuickPromoteEligibilityTweetFields: !0,
                 ...(0, o.d)(_),
               },
-              H
+              V
             )
             .then((e) => {
               var _;
-              return (null == (_ = e.timeline) ? void 0 : _.timeline) || L.cY;
+              return (null == (_ = e.timeline) ? void 0 : _.timeline) || U.cY;
             }),
       });
     },
     49890: (e, _, t) => {
       "use strict";
-      t.d(_, { B: () => a, Z: () => l });
+      t.d(_, { B: () => r, Z: () => l });
       var n = t(6899),
         i = t(46395),
-        r = t(90650);
-      const a = new n.fK.Entity(
+        a = t(90650);
+      const r = new n.fK.Entity(
           "teams",
-          { user: r.Z },
+          { user: a.Z },
           {
-            idAttribute: ({ user: e }) => e[r.Z.idAttribute] || e,
+            idAttribute: ({ user: e }) => e[a.Z.idAttribute] || e,
             processStrategy: i.Z,
           }
         ),
-        l = new n.fK.Array(a);
+        l = new n.fK.Array(r);
     },
     32091: (e, _, t) => {
       "use strict";
       t.d(_, { Gm: () => o, ZH: () => s, kH: () => l });
       var n = t(78772),
         i = t(52612),
-        r = t(57659),
-        a = t(62104);
+        a = t(57659),
+        r = t(62104);
       const l = { All: "any", AutoBlock: "autoblock", Imported: "imported" },
         o = (e) =>
-          (0, a.Z)({
+          (0, r.Z)({
             timelineId: `blockedAccountsGraphQL-${e}`,
             network: {
               getEndpoint: (_) =>
@@ -712,7 +768,7 @@
               .filter((e) => {
                 if (!e.content || e.type !== i.ZP.User) return !1;
                 const _ = e.content.id,
-                  t = r.ZP.select(n, _);
+                  t = a.ZP.select(n, _);
                 return t && !t.blocking;
               })
               .map((_) => e.removeEntry(_.entryId))
@@ -721,4 +777,4 @@
     },
   },
 ]);
-//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.Graphql.ec27dd5a.js.map
+//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.AccountTaxonomy.27bb991a.js.map
