+++ info/endpoints/endpoints.Collections.6dd079ca.js
@@ -48,4 +48,4 @@
     },
   },
 ]);
-//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.Collections.1a03cd6a.js.map
+//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.Collections.6dd079ca.js.map
