+++ info/endpoints/endpoints.Lists.530c138a.js
@@ -4,7 +4,7 @@
   {
     29848: (e) => {
       e.exports = {
-        queryId: "VeIlDt0K1vcCzf-jW4qd6A",
+        queryId: "h52d1F7dumWGE1tJAhQBpg",
         operationName: "BlockedAccountsAll",
         operationType: "query",
         metadata: {
@@ -34,7 +34,7 @@
     },
     57274: (e) => {
       e.exports = {
-        queryId: "h9d1UQkdRub8jR5XVY_uuQ",
+        queryId: "8w-D2OhT0jmGzXaNY--UQA",
         operationName: "BlockedAccountsAutoBlock",
         operationType: "query",
         metadata: {
@@ -64,7 +64,7 @@
     },
     2216: (e) => {
       e.exports = {
-        queryId: "gZpIitopgjdaqdwK6reCtg",
+        queryId: "8LDNeOEm0kA98uoDsqXvMg",
         operationName: "BlockedAccountsImported",
         operationType: "query",
         metadata: {
@@ -94,7 +94,7 @@
     },
     31354: (e) => {
       e.exports = {
-        queryId: "0j3btNWBmBQ6HSXNlruInw",
+        queryId: "rIxum3avpCu7APi7mxTNjw",
         operationName: "CombinedLists",
         operationType: "query",
         metadata: {
@@ -124,7 +124,7 @@
     },
     41839: (e) => {
       e.exports = {
-        queryId: "oW2F-EMg_7rtTJgDShBe5w",
+        queryId: "JpFFCTBPxYVlDqMUr9twzQ",
         operationName: "Followers",
         operationType: "query",
         metadata: {
@@ -154,7 +154,7 @@
     },
     97621: (e) => {
       e.exports = {
-        queryId: "1go5V_4KX2r8NMJgmPkhww",
+        queryId: "RvojYJJB90VwJ0rdVhbjMQ",
         operationName: "FollowersYouKnow",
         operationType: "query",
         metadata: {
@@ -184,7 +184,7 @@
     },
     87669: (e) => {
       e.exports = {
-        queryId: "RsIOCb3LhS_AzOkBxJiJIg",
+        queryId: "b22I8WSfQ8H4Ev8486xAlQ",
         operationName: "Following",
         operationType: "query",
         metadata: {
@@ -214,7 +214,7 @@
     },
     75840: (e) => {
       e.exports = {
-        queryId: "bnQ1HNHhBA4S_iB1SaT9HA",
+        queryId: "LZfAdxTdNolKXw6ZkoY_kA",
         operationName: "GenericTimelineById",
         operationType: "query",
         metadata: {
@@ -324,7 +324,7 @@
     },
     73825: (e) => {
       e.exports = {
-        queryId: "vvs9UquTbJ50eXanbLvpRw",
+        queryId: "Zf8ZwG57EKtss-rPlryIqg",
         operationName: "ListCreationRecommendedUsers",
         operationType: "query",
         metadata: {
@@ -362,7 +362,7 @@
     },
     67264: (e) => {
       e.exports = {
-        queryId: "DjZYZN_Vq4NeE3IM5TeKPg",
+        queryId: "-F4wsOirYNXjjg-ZjccQpQ",
         operationName: "ListEditRecommendedUsers",
         operationType: "query",
         metadata: {
@@ -392,7 +392,7 @@
     },
     4904: (e) => {
       e.exports = {
-        queryId: "VEwO8c1TYdm5zjxCOIiwFw",
+        queryId: "2TemLyqrMpTeAmysdbnVqw",
         operationName: "ListLatestTweetsTimeline",
         operationType: "query",
         metadata: {
@@ -422,7 +422,7 @@
     },
     77705: (e) => {
       e.exports = {
-        queryId: "a1BkkHLJhfT_2wVzGWx1KA",
+        queryId: "vA952kfgGw6hh8KatWnbqw",
         operationName: "ListMembers",
         operationType: "query",
         metadata: {
@@ -452,7 +452,7 @@
     },
     89234: (e) => {
       e.exports = {
-        queryId: "rHfuonnrX44LNbCAcr1btw",
+        queryId: "BlEXXdARdSeL_0KyKHHvvg",
         operationName: "ListMemberships",
         operationType: "query",
         metadata: {
@@ -490,7 +490,7 @@
     },
     67833: (e) => {
       e.exports = {
-        queryId: "rKmqARYlRBNR4IBm0_Bp9A",
+        queryId: "wQcOSjSQ8NtgxIwvYl1lMg",
         operationName: "ListOwnerships",
         operationType: "query",
         metadata: {
@@ -568,7 +568,7 @@
     },
     29773: (e) => {
       e.exports = {
-        queryId: "R_ySXF1FDL9NFmGUA5c5cA",
+        queryId: "07lytXX9oG9uCld1RY4b0w",
         operationName: "ListRankedTweetsTimeline",
         operationType: "query",
         metadata: {
@@ -630,7 +630,7 @@
     },
     17392: (e) => {
       e.exports = {
-        queryId: "hXUJLF9VH5xRZ4Iu0VJUOw",
+        queryId: "e57wIELAAe0fYt4Hmqsk6g",
         operationName: "ListSubscribers",
         operationType: "query",
         metadata: {
@@ -716,7 +716,7 @@
     },
     19227: (e) => {
       e.exports = {
-        queryId: "3BsvJLycU3b4f8U3u-hKfQ",
+        queryId: "ehnzbxPHA69pyaV2EydN1g",
         operationName: "ListsDiscovery",
         operationType: "query",
         metadata: {
@@ -746,7 +746,7 @@
     },
     63884: (e) => {
       e.exports = {
-        queryId: "GNNeFeFp6VNuti13-Ej6bQ",
+        queryId: "nhYp4n09Hi5n2hQWseQztg",
         operationName: "ListsManagementPageTimeline",
         operationType: "query",
         metadata: {
@@ -792,7 +792,7 @@
     },
     50931: (e) => {
       e.exports = {
-        queryId: "p3kCiX2fbX2lRjYgo6h6Ew",
+        queryId: "hnaqw2Vok5OETdBVa_uexw",
         operationName: "ModeratedTimeline",
         operationType: "query",
         metadata: {
@@ -822,7 +822,7 @@
     },
     59113: (e) => {
       e.exports = {
-        queryId: "003287_lTwFeSs_lOSw-HA",
+        queryId: "-G9eXTmseyiSenbqjrEG6w",
         operationName: "MutedAccounts",
         operationType: "query",
         metadata: {
@@ -852,7 +852,7 @@
     },
     26491: (e) => {
       e.exports = {
-        queryId: "-UBJTrL7wLKQXV-x9-bVZg",
+        queryId: "o0YtPFnd4Lk_pOQb9alCvA",
         operationName: "SuperFollowers",
         operationType: "query",
         metadata: {
@@ -901,14 +901,14 @@
       t.r(_),
         t.d(_, {
           default: () => O,
-          errorBreadcrumb: () => z,
-          isFatalGenericTimelineError: () => C,
-          isFatalModeratedTimelineError: () => M,
-          isFatalTeamsResponseError: () => U,
+          errorBreadcrumb: () => N,
+          isFatalGenericTimelineError: () => Y,
+          isFatalModeratedTimelineError: () => C,
+          isFatalTeamsResponseError: () => E,
           isFatalUserScopedFollowTimelineError: () => j,
-          isFatalViewerScopedFollowTimelineError: () => E,
-          parseUserScopedTimeline: () => A,
-          parseViewerScopedTimeline: () => P,
+          isFatalViewerScopedFollowTimelineError: () => U,
+          parseUserScopedTimeline: () => Z,
+          parseViewerScopedTimeline: () => z,
         });
       var i = t(58955),
         n = t(79363),
@@ -936,9 +936,9 @@
       var b = t(29848),
         u = t.n(b),
         c = t(57274),
-        w = t.n(c),
-        m = t(2216),
-        h = t.n(m),
+        m = t.n(c),
+        w = t(2216),
+        h = t.n(w),
         v = t(41839),
         g = t.n(v),
         f = t(97621),
@@ -949,16 +949,16 @@
         I = t.n(L),
         T = t(50931),
         S = t.n(T),
-        F = t(59113),
-        k = t.n(F),
+        k = t(59113),
+        F = t.n(k),
         Q = t(26491),
         B = t.n(Q),
-        N = t(16221),
-        G = t.n(N),
-        Z = t(82249);
-      const A = (e) => {
+        G = t(16221),
+        A = t.n(G),
+        P = t(82249);
+      const Z = (e) => {
           var _;
-          let t = Z.cY;
+          let t = P.cY;
           var i;
           null != e &&
             null != (_ = e.user) &&
@@ -966,23 +966,23 @@
             "User" === e.user.result.__typename &&
             (t =
               (null == (i = e.user.result.timeline) ? void 0 : i.timeline) ||
-              Z.cY);
+              P.cY);
           return t;
         },
-        z = (e) => {
+        N = (e) => {
           var _;
           return { code: null == (_ = e[0]) ? void 0 : _.code };
         },
-        P = (e) => {
-          let _ = Z.cY;
+        z = (e) => {
+          let _ = P.cY;
           var t;
           null != e &&
             e.viewer &&
             (_ =
-              (null == (t = e.viewer.timeline) ? void 0 : t.timeline) || Z.cY);
+              (null == (t = e.viewer.timeline) ? void 0 : t.timeline) || P.cY);
           return _;
         },
-        U = (e) =>
+        E = (e) =>
           Array.isArray(e) &&
           !(0, n.Z)(
             e,
@@ -1000,23 +1000,23 @@
           return (
             n ||
               (0, a.ZP)("GQL URT: Failed to render Follow timeline", {
-                extra: z(e),
+                extra: N(e),
               }),
             !n && (0, s.jB)(e)
           );
         },
-        E = (e, _) => {
+        U = (e, _) => {
           var t;
           const i = null == _ || null == (t = _.viewer) ? void 0 : t.timeline;
           return (
             i ||
               (0, a.ZP)("GQL URT: Failed to render Subscriptions timeline", {
-                extra: z(e),
+                extra: N(e),
               }),
             !i && (0, s.jB)(e)
           );
         },
-        R = (e, _) => {
+        M = (e, _) => {
           var t, i;
           const n =
             null == _ ||
@@ -1027,12 +1027,12 @@
           return (
             n ||
               (0, a.ZP)("GQL URT: Failed to render Muting timeline", {
-                extra: z(e),
+                extra: N(e),
               }),
             !n && (0, s.jB)(e)
           );
         },
-        M = (e, _) => {
+        C = (e, _) => {
           var t, i;
           const n =
             "Tweet" ===
@@ -1046,23 +1046,23 @@
             n ||
               (0, a.ZP)(
                 "GQL URT: Failed to render ModeratedTimeline timeline",
-                { extra: z(e) }
+                { extra: N(e) }
               ),
             !n && (0, s.jB)(e)
           );
         },
-        C = (e, _) => {
+        Y = (e, _) => {
           var t;
           const i = null == _ || null == (t = _.timeline) ? void 0 : t.timeline;
           return (
             i ||
               (0, a.ZP)("GQL URT: Failed to render GenericTimelineById", {
-                extra: z(e),
+                extra: N(e),
               }),
             !i && (0, s.jB)(e)
           );
         };
-      function Y(e, _, t) {
+      function R(e, _, t) {
         let i;
         function n(e) {
           return (
@@ -1074,7 +1074,7 @@
           );
         }
         return (
-          (i = e === r.kH.Imported ? h() : e === r.kH.AutoBlock ? w() : u()),
+          (i = e === r.kH.Imported ? h() : e === r.kH.AutoBlock ? m() : u()),
           ({ count: r, cursor: o }) =>
             _.graphQL(
               i,
@@ -1093,18 +1093,18 @@
                   i ||
                     (0, a.ZP)(
                       `GQL URT: Failed to query for Blocked Accounts [${e}]`,
-                      { extra: z(_) }
+                      { extra: N(_) }
                     ),
                   !i && (0, s.jB)(_)
                 );
               }
-            ).then((e) => n(e) || Z.cY)
+            ).then((e) => n(e) || P.cY)
         );
       }
       const O = ({ apiClient: e, featureSwitches: _ }) => ({
-        fetchBlockedAccountsImportedGraphql: Y(r.kH.Imported, e, _),
-        fetchBlockedAccountsAllGraphql: Y(r.kH.All, e, _),
-        fetchBlockedAccountsAutoBlockGraphql: Y(r.kH.AutoBlock, e, _),
+        fetchBlockedAccountsImportedGraphql: R(r.kH.Imported, e, _),
+        fetchBlockedAccountsAllGraphql: R(r.kH.All, e, _),
+        fetchBlockedAccountsAutoBlockGraphql: R(r.kH.AutoBlock, e, _),
         fetchFollowers: ({ count: t, cursor: i, userId: n }) =>
           e
             .graphQL(
@@ -1118,7 +1118,7 @@
               },
               j
             )
-            .then(A),
+            .then(Z),
         fetchFollowersYouKnow: ({ count: t, cursor: i, userId: n }) =>
           e
             .graphQL(
@@ -1132,7 +1132,7 @@
               },
               j
             )
-            .then(A),
+            .then(Z),
         fetchFollowing: ({ count: t, cursor: i, userId: n }) =>
           e
             .graphQL(
@@ -1146,7 +1146,7 @@
               },
               j
             )
-            .then(A),
+            .then(Z),
         fetchSuperFollowers: ({ count: t, cursor: i }) =>
           e
             .graphQL(
@@ -1157,12 +1157,12 @@
                 includePromotedContent: !1,
                 ...(0, l.d)(_),
               },
-              E
+              U
             )
-            .then(P),
+            .then(z),
         fetchViewerTeams: () =>
           e
-            .graphQLFullResponse(G(), { ...(0, l.S)(_) }, U)
+            .graphQLFullResponse(A(), { ...(0, l.S)(_) }, E)
             .then(({ data: e, errors: _ }) => {
               const { viewer: t } = e,
                 n = (t && t.teams && t.teams.act_as) || [];
@@ -1182,7 +1182,7 @@
                 includePromotedContent: !1,
                 ...(0, l.d)(_),
               },
-              M
+              C
             )
             .then((e) => {
               var _;
@@ -1195,29 +1195,29 @@
                   (null == (t = e.tweet.result) ||
                   null == (i = t.timeline_response)
                     ? void 0
-                    : i.timeline) || Z.cY
+                    : i.timeline) || P.cY
                 );
               }
-              return Z.cY;
+              return P.cY;
             }),
         fetchMutedAccounts: ({ count: t, cursor: i }) =>
           e
             .graphQL(
-              k(),
+              F(),
               {
                 count: t,
                 cursor: i,
                 includePromotedContent: !1,
                 ...(0, l.d)(_),
               },
-              R
+              M
             )
             .then((e) => {
               var _, t;
               return (
                 (null == (_ = e.viewer) || null == (t = _.muting_timeline)
                   ? void 0
-                  : t.timeline) || Z.cY
+                  : t.timeline) || P.cY
               );
             }),
         fetchGenericTimelineById: ({ count: t, cursor: i, timelineId: n }) =>
@@ -1231,11 +1231,11 @@
                 withQuickPromoteEligibilityTweetFields: !0,
                 ...(0, l.d)(_),
               },
-              C
+              Y
             )
             .then((e) => {
               var _;
-              return (null == (_ = e.timeline) ? void 0 : _.timeline) || Z.cY;
+              return (null == (_ = e.timeline) ? void 0 : _.timeline) || P.cY;
             }),
       });
     },
@@ -1243,15 +1243,15 @@
       "use strict";
       t.r(_),
         t.d(_, {
-          default: () => Ze,
+          default: () => Pe,
           isFatalAddMemberError: () => qe,
           isFatalFetchListByIdError: () => Se,
-          isFatalFetchListBySlugError: () => Fe,
+          isFatalFetchListBySlugError: () => ke,
           isFatalFetchMembersTimelineError: () => Be,
-          isFatalFetchPinsError: () => Ge,
-          isFatalFetchRecommendedUsersTimelineError: () => Ne,
+          isFatalFetchPinsError: () => Ae,
+          isFatalFetchRecommendedUsersTimelineError: () => Ge,
           isFatalFetchSubscribersTimelineError: () => Qe,
-          isFatalFetchTweetsTimelineError: () => ke,
+          isFatalFetchTweetsTimelineError: () => Fe,
           isFatalListsDiscoveryTimelineError: () => ve,
           isFatalListsManagementTimelineError: () => ge,
           isFatalListsTimelineError: () => he,
@@ -1275,8 +1275,8 @@
         b = t.n(p),
         u = t(42215),
         c = t.n(u),
-        w = t(4705),
-        m = t.n(w),
+        m = t(4705),
+        w = t.n(m),
         h = t(46405),
         v = t.n(h),
         g = t(3180),
@@ -1288,31 +1288,31 @@
         I = t(67264),
         T = t.n(I),
         S = t(4904),
-        F = t.n(S),
-        k = t(77705),
-        Q = t.n(k),
+        k = t.n(S),
+        F = t(77705),
+        Q = t.n(F),
         B = t(89234),
-        N = t.n(B),
-        G = t(27196),
-        Z = t.n(G),
-        A = t(67833),
-        z = t.n(A),
-        P = t(60935),
-        U = t.n(P),
+        G = t.n(B),
+        A = t(27196),
+        P = t.n(A),
+        Z = t(67833),
+        N = t.n(Z),
+        z = t(60935),
+        E = t.n(z),
         j = t(39610),
-        E = t.n(j),
-        R = t(18542),
-        M = t.n(R),
-        C = t(29773),
-        Y = t.n(C),
+        U = t.n(j),
+        M = t(18542),
+        C = t.n(M),
+        Y = t(29773),
+        R = t.n(Y),
         O = t(34105),
         V = t.n(O),
-        H = t(19227),
-        D = t.n(H),
-        J = t(63884),
-        X = t.n(J),
-        K = t(86526),
-        W = t.n(K),
+        D = t(19227),
+        H = t.n(D),
+        X = t(63884),
+        K = t.n(X),
+        J = t(86526),
+        W = t.n(J),
         $ = t(33309),
         ee = t.n($),
         _e = t(17392),
@@ -1329,9 +1329,9 @@
         be = t(82249),
         ue = t(97534),
         ce = t(87233),
-        we = t(55371);
-      const me = (e) => ({
-          ...we.getGlobalDefaults(e),
+        me = t(55371);
+      const we = (e) => ({
+          ...me.getGlobalDefaults(e),
           ...ce.getGlobalDefaults(e),
           cursor: -1,
         }),
@@ -1434,7 +1434,7 @@
             !t && (0, r.jB)(e)
           );
         },
-        Fe = (e, _) => {
+        ke = (e, _) => {
           var t;
           const i =
             null == _ || null == (t = _.user_by_screen_name) ? void 0 : t.list;
@@ -1443,7 +1443,7 @@
             !i && (0, r.jB)(e)
           );
         },
-        ke = (e, _) => {
+        Fe = (e, _) => {
           var t, i;
           const a =
             null == _ || null == (t = _.list) || null == (i = t.tweets_timeline)
@@ -1480,7 +1480,7 @@
             !a && (0, r.jB)(e)
           );
         },
-        Ne = (e, _) => {
+        Ge = (e, _) => {
           var t, i;
           const a =
             null == _ ||
@@ -1494,14 +1494,14 @@
             !a && (0, r.jB)(e)
           );
         },
-        Ge = (e, _) => {
+        Ae = (e, _) => {
           const t = null == _ ? void 0 : _.viewer;
           return (
             t || (0, n.ZP)("GQL List: Failed to fetch List Pins"),
             !t && (0, r.jB)(e)
           );
         },
-        Ze = ({ apiClient: e, featureSwitches: _ }) => ({
+        Pe = ({ apiClient: e, featureSwitches: _ }) => ({
           addToList(t, n) {
             const { list_id: a, user_id: r } = t;
             return e
@@ -1522,7 +1522,7 @@
           editBannerImage(t, n) {
             const { listId: a, mediaId: r } = t;
             return e
-              .graphQL(m(), { listId: a, mediaId: r, ...(0, s.S)(_) })
+              .graphQL(w(), { listId: a, mediaId: r, ...(0, s.S)(_) })
               .then((e) => (0, i.Fv)(e.list, l.Z));
           },
           deleteList(_, t) {
@@ -1538,7 +1538,7 @@
           fetchList(t, n) {
             if (t.list_id)
               return e
-                .graphQL(M(), { listId: t.list_id, ...(0, s.S)(_) }, Se)
+                .graphQL(C(), { listId: t.list_id, ...(0, s.S)(_) }, Se)
                 .then((e) => (0, i.Fv)(e.list, l.Z));
             if (t.slug) {
               const { owner_screen_name: n, slug: a } = t;
@@ -1546,7 +1546,7 @@
                 .graphQL(
                   v(),
                   { screenName: n, listSlug: a, ...(0, s.S)(_) },
-                  Fe
+                  ke
                 )
                 .then((e) => {
                   var _;
@@ -1575,7 +1575,7 @@
               .then(ue.parseUserScopedTimeline),
           fetchListsManagementPageTimeline: ({ count: t, cursor: i }) =>
             e
-              .graphQL(X(), { count: t, cursor: i, ...(0, s.d)(_) }, ge)
+              .graphQL(K(), { count: t, cursor: i, ...(0, s.d)(_) }, ge)
               .then((e) => {
                 var _, t;
                 return (
@@ -1587,9 +1587,9 @@
               }),
           fetchTweetsGraphQL(t, i) {
             const { count: n, cursor: a, listId: r, useRanked: l } = t,
-              o = l ? Y() : F(),
+              o = l ? R() : k(),
               d = { listId: r, count: n, cursor: a, ...(0, s.d)(_) };
-            return e.graphQL(o, d, ke).then((e) => {
+            return e.graphQL(o, d, Fe).then((e) => {
               var _;
               return (
                 (null == (_ = e.list.tweets_timeline) ? void 0 : _.timeline) ||
@@ -1626,7 +1626,7 @@
             const { count: n, cursor: r, displayLocation: l, listId: o } = t,
               d = l === a.O.Creation ? q() : T(),
               p = { listId: o, count: n, cursor: r, ...(0, s.d)(_) };
-            return e.graphQL(d, p, Ne).then((e) => {
+            return e.graphQL(d, p, Ge).then((e) => {
               var _;
               return (
                 (null == (_ = e.list.recommended_users)
@@ -1636,7 +1636,7 @@
             });
           },
           fetchPins: () =>
-            e.graphQL(E(), { ...(0, s.S)(_) }, Ge).then((e) => {
+            e.graphQL(U(), { ...(0, s.S)(_) }, Ae).then((e) => {
               var _;
               const t =
                 null == e || null == (_ = e.viewer) ? void 0 : _.pinned_lists;
@@ -1661,7 +1661,7 @@
           },
           fetchSuggestedLists: ({ count: t, cursor: i }) =>
             e
-              .graphQL(D(), { count: t, cursor: i, ...(0, s.d)(_) }, ve)
+              .graphQL(H(), { count: t, cursor: i, ...(0, s.d)(_) }, ve)
               .then((e) => {
                 var _, t;
                 return (
@@ -1675,7 +1675,7 @@
             const { count: n, cursor: a, targetUserId: r, userId: l } = t;
             return e
               .graphQL(
-                z(),
+                N(),
                 {
                   userId: l,
                   isListMemberTargetUserId: r,
@@ -1689,13 +1689,13 @@
           },
           fetchMemberships: (t, n) =>
             e
-              .get("lists/memberships", { ...me(_), ...t }, n || {})
+              .get("lists/memberships", { ...we(_), ...t }, n || {})
               .then((e) => (0, i.Fv)(e, pe.kD)),
           fetchMembershipsGraphQL(t, i) {
             const { count: n, cursor: a, userId: r } = t;
             return e
               .graphQL(
-                N(),
+                G(),
                 { userId: r, count: n, cursor: a, ...(0, s.d)(_) },
                 he
               )
@@ -1712,7 +1712,7 @@
               a = (0, s.S)(_);
             return i
               ? e.graphQL(re(), { listId: n, ...a }, Te)
-              : e.graphQL(U(), { listId: n, ...a }, Le);
+              : e.graphQL(E(), { listId: n, ...a }, Le);
           },
           pinManyLists(t) {
             const { listIds: n } = t;
@@ -1738,7 +1738,7 @@
           toggleMute(_, t) {
             const { listId: i, mute: n } = _;
             return n
-              ? e.graphQL(Z(), { listId: i })
+              ? e.graphQL(P(), { listId: i })
               : e.graphQL(ne(), { listId: i });
           },
           editList(t, n) {
@@ -1826,4 +1826,4 @@
     },
   },
 ]);
-//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.Lists.ac03d01a.js.map
+//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.Lists.530c138a.js.map
