+++ info/endpoints/endpoints.LiveEvent.6258c53a.js
@@ -64,4 +64,4 @@
     },
   },
 ]);
-//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.LiveEvent.fdf9ea6a.js.map
+//# sourceMappingURL=https://ton.local.twitter.com/responsive-web-internal/sourcemaps/client-web/endpoints.LiveEvent.6258c53a.js.map
