+++ twitter/config/operations.py
@@ -1441,39 +1441,35 @@ operations = {
         }
     },
     "HomeTimeline": {
-        "queryId": "u5XGoPEa9-Hb30zpGz3Yrg",
-        "variables": {
-            "count": 1000,
-            "includePromotedContent": True,
-            "latestControlAvailable": True,
-            "requestContext": "launch",
-            "withCommunity": True,
-            "withSuperFollowsUserFields": True,
-            "withDownvotePerspective": False,
-            "withReactionsMetadata": False,
-            "withReactionsPerspective": False,
-            "withSuperFollowsTweetFields": True
+        'queryId': 'HCosKfLNW1AcOo3la3mMgg',
+        'variables': {
+            'count': 1000,
+            'includePromotedContent': True,
+            'latestControlAvailable': True,
+            'requestContext': 'launch',
+            'withCommunity': True,
+            # 'seenTweetIds': [],
         },
-        "features": {
-            "responsive_web_twitter_blue_verified_badge_is_enabled": True,
-            "responsive_web_graphql_exclude_directive_enabled": True,
-            "verified_phone_label_enabled": True,
-            "responsive_web_graphql_timeline_navigation_enabled": True,
-            "responsive_web_graphql_skip_user_profile_image_extensions_enabled": True,
-            "tweetypie_unmention_optimization_enabled": True,
-            "vibe_api_enabled": True,
-            "responsive_web_edit_tweet_api_enabled": True,
-            "graphql_is_translatable_rweb_tweet_is_translatable_enabled": True,
-            "view_counts_everywhere_api_enabled": True,
-            "longform_notetweets_consumption_enabled": True,
-            "tweet_awards_web_tipping_enabled": True,
-            "freedom_of_speech_not_reach_fetch_enabled": True,
-            "standardized_nudges_misinfo": True,
-            "tweet_with_visibility_results_prefer_gql_limited_actions_policy_enabled": True,
-            "interactive_text_enabled": True,
-            "responsive_web_text_conversations_enabled": True,
-            "longform_notetweets_richtext_consumption_enabled": True,
-            "responsive_web_enhance_cards_enabled": True
+        'features': {
+            'blue_business_profile_image_shape_enabled': True,
+            'responsive_web_graphql_exclude_directive_enabled': True,
+            'verified_phone_label_enabled': False,
+            'responsive_web_graphql_timeline_navigation_enabled': True,
+            'responsive_web_graphql_skip_user_profile_image_extensions_enabled': False,
+            'tweetypie_unmention_optimization_enabled': True,
+            'vibe_api_enabled': True,
+            'responsive_web_edit_tweet_api_enabled': True,
+            'graphql_is_translatable_rweb_tweet_is_translatable_enabled': True,
+            'view_counts_everywhere_api_enabled': True,
+            'longform_notetweets_consumption_enabled': True,
+            'tweet_awards_web_tipping_enabled': False,
+            'freedom_of_speech_not_reach_fetch_enabled': False,
+            'standardized_nudges_misinfo': True,
+            'tweet_with_visibility_results_prefer_gql_limited_actions_policy_enabled': False,
+            'interactive_text_enabled': True,
+            'responsive_web_text_conversations_enabled': False,
+            'longform_notetweets_rich_text_read_enabled': True,
+            'responsive_web_enhance_cards_enabled': False,
         }
     },
     "ImmersiveMedia": {
